Simplifies PcrSlot and moves it to its own file.
This commit moves the PcrSlot enum to its own source file
and simplifies the conversion from a u32.
