+++ tss-esapi/src/structures/pcr/mod.rs
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 pub mod select;
 pub mod selection;
+pub mod slot;
